requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
qrcode[pil]==7.4.2
Pillow==10.3.0
jinja2==3.1.4
python-multipart==0.0.9
websockets==12.0
aiosqlite==0.20.0
